Delete CNAME
The waldyrious.net domain is currently set up to use an [A record pointing the github servers][1], so the CNAME record should be redundant. It would be useful to [set up a www redirect][2], but that's not a deal-breaker, so let's remove it for now.

Hopefully this might help with the [recently announced HTTPS support][3], which is currently working for https://waldyrious.github.io but not for https://waldyrious.net/ at the moment (I'm getting an "Insecure Connection" error, saying "waldyrious.net uses an invalid security certificate. The certificate is only valid for the following names: *.github.com, github.com, *.github.io, github.io. Error code: SSL_ERROR_BAD_CERT_DOMAIN")

[1]: https://help.github.com/articles/setting-up-an-apex-domain/
[2]: https://help.github.com/articles/setting-up-an-apex-domain-and-www-subdomain/
[3]: https://blog.github.com/2018-05-01-github-pages-custom-domains-https/