fix visited link color in print version of pdf
The goal of color:inherit was to set the color of visited links to match the non-visited ones
but it ended up inheriting from the parent element instead (black).
It doesn't seem to be possible to do this with CSS alone
(see https://stackoverflow.com/q/1929618/266309).
color:initial and color:unset don't work either.

So the workaround is to define the same link color present in downstyler
both for regular links and for visited links in the print version of the cv page.

This change is a partial revert of 2950480a2 and 41db90e24.